# api/internal/httpapi/static/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Deadlock Patch Notes API",
    "version": "1.1.0",
    "description": "HTTP API for patch timelines, heroes, items, and spells used by deadlock.jakubdolenek.xyz."
  },
  "servers": [
    {
      "url": "https://deadlock.jakubdolenek.xyz/api",
      "description": "Production"
    }
  ],
...
  "paths": {
...
    "/v1/patches": {
      "get": {
        "tags": [
          "Patches"
        ],
        "summary": "List patches",
        "operationId": "listPatches",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "description": "Page number (default 1). Invalid values fall back to 1.",
            "schema": {
              "type": "integer",
              "default": 1,
              "minimum": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "description": "Page size (default 12, max 50). Invalid values fall back to 12.",
            "schema": {
              "type": "integer",
              "default": 12,
              "minimum": 1,
              "maximum": 50
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Patch list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatchListResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/patches/{slug}": {
      "get": {
        "tags": [
          "Patches"
        ],
        "summary": "Get patch detail",
        "operationId": "getPatchBySlug",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Patch detail",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatchDetail"
                }
              }
            }
          },
          "404": {
            "description": "Patch not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/heroes": {
      "get": {
        "tags": [
          "Heroes"
        ],
        "summary": "List heroes",
        "operationId": "listHeroes",
        "responses": {
          "200": {
            "description": "Hero list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HeroListResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/heroes/{heroSlug}/changes": {
      "get": {
        "tags": [
          "Heroes"
        ],
        "summary": "Get hero timeline",
        "operationId": "getHeroChanges",
        "parameters": [
          {
            "name": "heroSlug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "skill",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Skill name filter. Use `general` for general-only entries."
          },
          {
            "name": "from",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Inclusive UTC start. Accepts RFC3339 or YYYY-MM-DD."
          },
          {
            "name": "to",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Inclusive UTC end. Accepts RFC3339 or YYYY-MM-DD."
          }
        ],
        "responses": {
          "200": {
            "description": "Hero timeline",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HeroChangesResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid query parameter",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Hero not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/items": {
      "get": {
        "tags": [
          "Items"
        ],
        "summary": "List items",
        "operationId": "listItems",
        "responses": {
          "200": {
            "description": "Item list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemListResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/items/{itemSlug}/changes": {
      "get": {
        "tags": [
          "Items"
        ],
        "summary": "Get item timeline",
        "operationId": "getItemChanges",
        "parameters": [
          {
            "name": "itemSlug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "from",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Inclusive UTC start. Accepts RFC3339 or YYYY-MM-DD."
          },
          {
            "name": "to",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Inclusive UTC end. Accepts RFC3339 or YYYY-MM-DD."
          }
        ],
        "responses": {
          "200": {
            "description": "Item timeline",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemChangesResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid query parameter",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Item not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/spells": {
      "get": {
        "tags": [
          "Spells"
        ],
        "summary": "List spells",
        "operationId": "listSpells",
        "responses": {
          "200": {
            "description": "Spell list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpellListResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/spells/{spellSlug}/changes": {
      "get": {
        "tags": [
          "Spells"
        ],
        "summary": "Get spell timeline",
        "operationId": "getSpellChanges",
        "parameters": [
          {
            "name": "spellSlug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "from",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Inclusive UTC start. Accepts RFC3339 or YYYY-MM-DD."
          },
          {
            "name": "to",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Inclusive UTC end. Accepts RFC3339 or YYYY-MM-DD."
          }
        ],
        "responses": {
          "200": {
            "description": "Spell timeline",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpellChangesResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid query parameter",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Spell not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ErrorPayload": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "requestId": {
            "type": "string"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorPayload"
          }
        }
      },
      "PatchSource": {
        "type": "object",
        "required": [
          "type",
          "url"
        ],
        "properties": {
          "type": {
            "type": "string"
          },
          "url": {
            "type": "string"
          }
        }
      },
      "PatchChange": {
        "type": "object",
        "required": [
          "id",
          "text"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "text": {
            "type": "string"
          }
        }
      },
...
      "PatchEntry": {
        "type": "object",
        "required": [
          "id",
          "entityName",
          "changes"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "entityName": {
            "type": "string"
          },
          "entityIconUrl": {
            "type": "string"
          },
          "entityIconFallbackUrl": {
            "type": "string"
          },
          "summary": {
            "type": "string"
          },
          "changes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PatchChange"
            }
          },
          "groups": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PatchEntryGroup"
            }
          }
        }
      },
      "PatchSection": {
        "type": "object",
        "required": [
          "id",
          "title",
          "kind",
          "entries"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "kind": {
            "type": "string",
            "enum": [
              "general",
              "items",
              "heroes"
            ]
          },
          "entries": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PatchEntry"
            }
          }
        }
      },
      "PatchTimelineSummary": {
        "type": "object",
        "required": [
          "id",
          "releaseType",
          "title",
          "releasedAt"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "releaseType": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "releasedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "PatchSummary": {
        "type": "object",
        "required": [
          "id",
          "slug",
          "title",
          "publishedAt",
          "category",
          "imageUrl",
          "source"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "publishedAt": {
            "type": "string",
            "format": "date-time"
          },
          "category": {
            "type": "string"
          },
          "imageUrl": {
            "type": "string"
          },
          "source": {
            "$ref": "#/components/schemas/PatchSource"
          },
          "releaseTimeline": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PatchTimelineSummary"
            }
          }
        }
      },
      "Pagination": {
        "type": "object",
        "required": [
          "page",
          "pageSize",
          "totalItems",
          "totalPages"
        ],
        "properties": {
          "page": {
            "type": "integer"
          },
          "pageSize": {
            "type": "integer"
          },
          "totalItems": {
            "type": "integer"
          },
          "totalPages": {
            "type": "integer"
          }
        }
      },
      "PatchListResponse": {
        "type": "object",
        "required": [
          "patches",
          "pagination"
        ],
        "properties": {
          "patches": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PatchSummary"
            }
          },
          "pagination": {
            "$ref": "#/components/schemas/Pagination"
          }
        }
      },
      "PatchTimelineBlock": {
        "type": "object",
        "required": [
          "id",
          "releaseType",
          "title",
          "releasedAt",
          "source",
          "changes"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "releaseType": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "releasedAt": {
            "type": "string",
            "format": "date-time"
          },
          "source": {
            "$ref": "#/components/schemas/PatchSource"
          },
          "changes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PatchChange"
            }
          },
          "sections": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PatchSection"
            }
          }
        }
      },
      "PatchDetail": {
        "type": "object",
        "required": [
          "id",
          "slug",
          "title",
          "publishedAt",
          "category",
          "source",
          "imageUrl",
          "intro",
          "sections"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "publishedAt": {
            "type": "string",
            "format": "date-time"
          },
          "category": {
            "type": "string"
          },
          "source": {
            "$ref": "#/components/schemas/PatchSource"
          },
          "imageUrl": {
            "type": "string"
          },
          "intro": {
            "type": "string"
          },
          "sections": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PatchSection"
            }
          },
          "releaseTimeline": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PatchTimelineBlock"
            }
          }
        }
      },
      "HeroSummary": {
        "type": "object",
        "required": [
          "slug",
          "name",
          "lastChangedAt"
        ],
        "properties": {
          "slug": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "iconUrl": {
            "type": "string"
          },
          "iconFallbackUrl": {
            "type": "string"
          },
          "lastChangedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "HeroListResponse": {
        "type": "object",
        "required": [
          "heroes"
        ],
        "properties": {
          "heroes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/HeroSummary"
            }
          }
        }
      },
      "TimelinePatchRef": {
        "type": "object",
        "required": [
          "slug",
          "title"
        ],
        "properties": {
          "slug": {
            "type": "string"
          },
          "title": {
            "type": "string"
          }
        }
      },
      "HeroTimelineSkill": {
        "type": "object",
        "required": [
          "id",
          "title",
          "changes"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "iconUrl": {
            "type": "string"
          },
          "iconFallbackUrl": {
            "type": "string"
          },
          "changes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PatchChange"
            }
          }
        }
      },
      "HeroTimelineBlock": {
        "type": "object",
        "required": [
          "id",
          "releaseType",
          "displayLabel",
          "releasedAt",
          "patchRef",
          "source",
          "skills"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "releaseType": {
            "type": "string"
          },
          "displayLabel": {
            "type": "string"
          },
          "releasedAt": {
            "type": "string",
            "format": "date-time"
          },
          "patchRef": {
            "$ref": "#/components/schemas/TimelinePatchRef"
          },
          "source": {
            "$ref": "#/components/schemas/PatchSource"
          },
          "generalChanges": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PatchChange"
            }
          },
          "skills": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/HeroTimelineSkill"
            }
          }
        }
      },
      "HeroChangesResponse": {
        "type": "object",
        "required": [
          "hero",
          "timeline"
        ],
        "properties": {
          "hero": {
            "$ref": "#/components/schemas/HeroSummary"
          },
          "timeline": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/HeroTimelineBlock"
            }
          }
        }
      },
      "ItemSummary": {
        "type": "object",
        "required": [
          "slug",
          "name",
          "lastChangedAt"
        ],
        "properties": {
          "slug": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "iconUrl": {
            "type": "string"
          },
          "iconFallbackUrl": {
            "type": "string"
          },
          "lastChangedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ItemListResponse": {
        "type": "object",
        "required": [
          "items"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ItemSummary"
            }
          }
        }
      },
      "ItemTimelineBlock": {
        "type": "object",
        "required": [
          "id",
          "releaseType",
          "displayLabel",
          "releasedAt",
          "patchRef",
          "source",
          "changes"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "releaseType": {
            "type": "string"
          },
          "displayLabel": {
            "type": "string"
          },
          "releasedAt": {
            "type": "string",
            "format": "date-time"
          },
          "patchRef": {
            "$ref": "#/components/schemas/TimelinePatchRef"
          },
          "source": {
            "$ref": "#/components/schemas/PatchSource"
          },
          "changes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PatchChange"
            }
          }
        }
      },
      "ItemChangesResponse": {
        "type": "object",
        "required": [
          "item",
          "timeline"
        ],
        "properties": {
          "item": {
            "$ref": "#/components/schemas/ItemSummary"
          },
          "timeline": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ItemTimelineBlock"
            }
          }
        }
      },
      "SpellSummary": {
        "type": "object",
        "required": [
          "slug",
          "name",
          "lastChangedAt"
        ],
        "properties": {
          "slug": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "iconUrl": {
            "type": "string"
          },
          "iconFallbackUrl": {
            "type": "string"
          },
          "lastChangedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "SpellListResponse": {
        "type": "object",
        "required": [
          "spells"
        ],
        "properties": {
          "spells": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SpellSummary"
            }
          }
        }
      },
      "SpellTimelineEntry": {
        "type": "object",
        "required": [
          "id",
          "changes"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "heroSlug": {
            "type": "string"
          },
          "heroName": {
            "type": "string"
          },
          "heroIconUrl": {
            "type": "string"
          },
          "heroIconFallbackUrl": {
            "type": "string"
          },
          "changes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PatchChange"
            }
          }
        }
      },
      "SpellTimelineBlock": {
        "type": "object",
        "required": [
          "id",
          "releaseType",
          "displayLabel",
          "releasedAt",
          "patchRef",
          "source",
          "entries"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "releaseType": {
            "type": "string"
          },
          "displayLabel": {
            "type": "string"
          },
          "releasedAt": {
            "type": "string",
            "format": "date-time"
          },
          "patchRef": {
            "$ref": "#/components/schemas/TimelinePatchRef"
          },
          "source": {
            "$ref": "#/components/schemas/PatchSource"
          },
          "entries": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SpellTimelineEntry"
            }
          }
        }
      },
      "SpellChangesResponse": {
        "type": "object",
        "required": [
          "spell",
          "timeline"
        ],
        "properties": {
          "spell": {
            "$ref": "#/components/schemas/SpellSummary"
          },
          "timeline": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SpellTimelineBlock"
            }
          }
        }
      }
    }
  }
}